document and add some tools

+++ mac_setup.sh
@@ -14,6 +14,9 @@
 # Safe Removal: Remove individual aliases by unsetting (e.g., unalias la).
 ###############################################################################
 
+# Provide a default color flag if not already defined by the environment/shell rc
+: "${colorflag:='--color=auto'}"
+
 # List all files colorized in long format, including dot files
 alias la="ls -laF ${colorflag}"
 
@@ -25,7 +28,8 @@ alias ls="command ls ${colorflag}"
 export LS_COLORS='no=00:fi=00:di=01;34:ln=01;36:pi=40;33:so=01;35:do=01;35:bd=40;33;01:cd=40;33;01:or=40;31;01:ex=01;32:*.tar=01;31:*.tgz=01;31:*.arj=01;31:*.taz=01;31:*.lzh=01;31:*.zip=01;31:*.z=01;31:*.Z=01;31:*.gz=01;31:*.bz2=01;31:*.deb=01;31:*.rpm=01;31:*.jar=01;31:*.jpg=01;35:*.jpeg=01;35:*.gif=01;35:*.bmp=01;35:*.pbm=01;35:*.pgm=01;35:*.ppm=01;35:*.tga=01;35:*.xbm=01;35:*.xpm=01;35:*.tif=01;35:*.tiff=01;35:*.png=01;35:*.mov=01;35:*.mpg=01;35:*.mpeg=01;35:*.avi=01;35:*.fli=01;35:*.gl=01;35:*.dl=01;35:*.xcf=01;35:*.xwd=01;35:*.ogg=01;35:*.mp3=01;35:*.wav=01;35:'
 
 # Get OS X Software Updates, and update installed Homebrew, and their installed packages
-alias update='sudo softwareupdate -i -a; brew update; brew upgrade --all; brew cleanup'
+# Resilient update alias (skips brew steps if brew not present)
+alias update='sudo softwareupdate -i -a; command -v brew >/dev/null 2>&1 && { brew update; brew upgrade; brew cleanup; }'
 
 # IP addresses
 alias ip="dig +short myip.opendns.com @resolver1.opendns.com"
@@ -89,4 +93,9 @@ alias ipcalculator='whatmask'
 alias ping='prettyping --nolegend'
 
 ###
-ln -s /opt/homebrew/opt/mysql-client@5.7/bin/mysql /opt/homebrew/bin/mysql
+# Guarded MySQL legacy symlink creation
+if [ -x /opt/homebrew/opt/mysql-client@5.7/bin/mysql ]; then
+	if [ ! -e /opt/homebrew/bin/mysql ]; then
+		ln -s /opt/homebrew/opt/mysql-client@5.7/bin/mysql /opt/homebrew/bin/mysql
+	fi
+fi
